Refactor folder structre
- Move all test code to a folder.
- Relocate TaksCollection to base package.
- Import necessary packages after refactor.
